#major modifications completed

+++ backend/src/main/java/com/ecommerce/nuance/repository/BookRepository.java
@@ -12,11 +12,11 @@ import org.springframework.stereotype.Repository;
 import com.ecommerce.nuance.model.Book;
 
 @Repository
-public interface BookRepository extends JpaRepository<Book, Integer>
+public interface BookRepository extends JpaRepository<Book, Long>
 {
 	List<Book> findByGenre(String genre);
-	Optional<Book> findById(Integer id);
+	Optional<Book> findById(long id);
 	List<Book> findByPriceLessThanEqual(double price);
 	List<Book> findByAuthorAndGenreIgnoreCase(String author, String genre);
-	Optional<Book> findByIdAndName(Long id, String bookName);
+	Optional<Book> findByIdAndName(long id, String bookName);
 }
